cmount: fix macOS hang: don't call host.Umount if a signal has been received
Before this change cgofuse and libatexit would race to see who could
unmount the file system with unpredicatable results. On macOS this
resulted in a hang on exit. On Linux it could report an error or not,
depending.

This change checks to see if umount is beng called from a signal and
if so leaves the unmounting to cgofuse/libfuse.

Fixes #4804